Sorting, types, logging
