+++ src/helpers/types/status.ts
@@ -1,0 +1,7 @@
+enum Status {
+  Success = "success",
+  Fail = "fail",
+  Error = "error",
+}
+
+export default Status;
